requirements.txt
aio-pika==6.8.0
google==3.0.0
protobuf==3.15.8
pytest==6.2.4
pytest-asyncio==0.15.1
pyzmq==22.0.3
